fix: b3 extract priority
